Added EER model for the database
Revised version from the one made in the meeting. Should be easier to read and is definitely easier to edit. It's a litter hard to believe this is the same information, so review it to be sure it's the same as the one we made in the meeting.

If there should be changes, I will make it, as I have the original layered file.